refactor: extract audio commands into ai-audio.ts (-891 lines from ai.ts)
Extracts 11 audio commands using the pattern established in ai-motion.ts:

ai-audio.ts (~941 lines):
  Commands: transcribe, tts, voices, sfx, isolate,
            voice-clone, music, music-status, audio-restore, dub, duck
  registerAudioCommands(cmd) — called from ai.ts

ai.ts:
  Imports registerAudioCommands from ai-audio.ts
  Shrinks from 12,760 → 11,869 lines (-891 net)

All 232 tests pass (51 repl tests were removed in #2).
